requirements.txt
absl-py==2.0.0
Chantlr4-python3-runtime==4.8
appdirs==1.4.4
asttokens==2.4.1
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
cloudpickle==3.0.0
cmake==3.27.7
contourpy==1.1.1
cycler==0.12.1
decorator==4.4.2
dm-control==1.0.15
dm-env==1.5
dm-tree==0.1.7
docker-pycreds==0.4.0
etils==1.3.0
exceptiongroup==1.1.3
executing==2.0.1
Farama-Notifications==0.0.4
filelock==3.13.0
fonttools==4.43.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
glfw==2.5.3
google-auth==2.23.4
google-auth-oauthlib==1.0.0
grpcio==1.59.0
gym-notices==0.0.8
gymnasium==0.29.1
hydra-core==1.1.0
hydra-submitit-launcher==1.1.5
idna==3.4
imageio==2.31.6
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipdb==0.13.9
ipython
isort==5.12.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
labmaze==1.0.6
lit==17.0.5
lxml==4.9.3
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
moviepy==1.0.3
mpmath==1.3.0
mujoco==3.0.1
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.1.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.0.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
Pillow==10.0.1
proglog==0.1.10
prompt-toolkit==3.0.41
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.17.1
PyOpenGL==3.1.7
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==1.32.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
stable-baselines3==2.1.0
stack-data==0.6.3
submitit==1.5.1
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==1.1.0
threadpoolctl==3.2.0
toml==0.10.2
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
traitlets==5.13.0
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wandb==0.15.12
wcwidth==0.2.10
Werkzeug==3.0.1
yapf==0.31.0
zipp==3.17.0
